KooSlurmInstallAutomationRefactory/dashboard/kooCAEWebServer_5000/requirements.txt
blinker==1.9.0
certifi==2025.7.9
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
Flask==3.1.1
flask-cors==6.0.1
Flask-SQLAlchemy==3.1.1
greenlet==3.2.3
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
numpy==2.3.1
numpy-stl==3.2.0
pybind11==3.0.0
python-dotenv==1.1.1
python-utils==3.9.1
PyYAML==6.0.1
requests==2.32.4
setuptools==80.9.0
SQLAlchemy==2.0.41
stl==0.0.3
trimesh==4.7.0
typing_extensions==4.14.1
urllib3==2.5.0
Werkzeug==3.1.3
